Setup channel to pass room ID into lighting shader
Plan is to override the RT0.a channel which is used for occlusion and
GI. If we do need occlusion maps, or GI, then we may need to find
another way, perhaps via the Specular Color channels.

Created RoomSurface shader which will be the material used for floors,
walls and ceilings, and a replacement for the standard shader used
currently.

Modified Room.cs->SetTexture to set the main texture on the material and
not necessarily create a new Material, instead create a material
instance.

Set up custom DeferredLighting shader as a replacement for Unity's
built-in, this way we can add in custom ambient lighting via the room
ID, and some sort of ambient color atlas.