Simplify approach to ambient light
Turns out it is simpler to simply apply ambient light as emission in the
shader, and set occlusion to 0, to prevent Unity from applying it's
ambient lighting.

The idea then is we simply pass the ambient color to the renderers on a
per-material basis. Then we multiply the Albedo by it, and call it a
day.