Editor script creates Moods enum when list changes
When the Moods list in LightingHelper changes, an editor script
generates a c# enum in LightingHelper based on the moods in the list.
This way when programmers want to set a mood, they can do so by enum
instead of string, and they can know via code which moods are available.
And to change that enum, one only needs to edit the list in the
inspector