Basics of rendering fog map
The idea is to have a top down orthographic camera that uses a
replacement shader and only renders the room floor meshes since the
ceilings are facing the other way. Currently it simply encodes the
moodIndex which is only 0 or 1, but later we will use that mood index to
sample fog colors and write those in.