Hack to prevent light bleed
The hack is basically that we associate an ID with each light and with
each room. Those ID's get compared in a custom deferred lighting shader
to determine if the lighting gets applied.

It's a hack because the ID gets placed into the specular color of the
materials which means they will all need to have the same specular
color. The ID for the lights gets packed into the alpha channel which
feels a little more benign.

Global lights will have alpha to 1, and materials not designed to
receive lighting from any of the rooms (like hallway etc.) should have
their specular red channel set to 1.